Move setup to ready() function
