+++ backend/gunicorn.conf.py
@@ -2,4 +2,4 @@ import multiprocessing
 
 bind = "0.0.0.0:8000"
 workers = 5
-timeout = 120
+timeout = 1200
